Fix #1548 about Solaris SIGBUS with accurate mode backup
The memory structure was not aligned properly on sparc systems (8 bytes)

+++ bacula/src/lib/htable.h
@@ -70,8 +70,8 @@ struct hlink {
 
 struct h_mem {
    struct h_mem *next;                /* next buffer */
-   char *mem;                         /* memory pointer */
-   int32_t rem;                       /* remaining bytes in big_buffer */
+   char   *mem;                       /* memory pointer */
+   int64_t rem;                       /* remaining bytes in big_buffer */
    char first[1];                     /* first byte */
 };
